upload script
grabs resume from 'grabs' folder. Only works if there is exactly 1 file in that folder. Breaks resume down, tokenizes, and turns it into a class probability distribution prediction for which categories the resume could fit into. Deletes file in 'grabs' folder when done. Displays an error if file deletion fails, but that case should never happen. Tested with a basic resume and is working as expected.